+++ hypnotoad/core/mesh.py
@@ -26,6 +26,7 @@ import numbers
 import re
 import sys
 import warnings
+import yaml
 
 import numpy
 from optionsfactory import OptionsFactory, WithMeta
@@ -3540,7 +3541,23 @@ class BoutMesh(Mesh):
             # looking at a different variable, which would be inconvenient. It is not
             # likely that we need to load the hypnotoad inputs in BOUT++, so no reason
             # to save as an attribute.
-            f.write("hypnotoad_inputs", self.equilibrium._getOptionsAsString())
+            #
+            # Some options may be duplicated because of saving both equilibrium
+            # and mesh options, but no other way to make sure to get everything
+            # from subclasses (i.e. TokamakEquilibrium)
+            inputs_string = (
+                "Equilibrium options:\n"
+                + self.equilibrium.user_options.as_table()
+                + "\nNonorthogonal equilibrium options:\n"
+                + self.equilibrium.nonorthogonal_options.as_table()
+                + "\nMesh options:\n"
+                + self.user_options.as_table()
+            )
+            f.write("hypnotoad_inputs", inputs_string)
+            options_dict = dict(self.equilibrium.user_options)
+            options_dict.update(self.equilibrium.nonorthogonal_options)
+            options_dict.update(self.user_options)
+            f.write("hypnotoad_inputs_yaml", yaml.dump(options_dict))
 
             f.write_file_attribute("hypnotoad_version", self.version)
             if self.git_hash is not None:
@@ -3599,6 +3616,3 @@ class BoutMesh(Mesh):
                 "Some variable has not been defined yet: have you called "
                 "Mesh.geometry()?"
             )
-
-    def saveOptions(self, filename="hypnotoad_options.yaml"):
-        self.equilibrium.saveOptions(filename)