+++ doc/whats-new.md
@@ -10,6 +10,9 @@ What's new
 - Fix position of start and end points of contours when refining and add more tolerance
   settings, to enable more reproducible grid generation (#95)\
   By [John Omotani](https://github.com/johnomotani)
+- Options are saved as a YAML string in "hypnotoad_inputs_yaml" to make them
+  easier to read in code later. (#98)\
+  By [John Omotani](https://github.com/johnomotani)
 - When exceptions are caught by the GUI, print the traceback as well as the
   exception message (#95)\
   By [John Omotani](https://github.com/johnomotani)
@@ -22,6 +25,9 @@ What's new
 
 ### Bug fixes
 
+- Save all options to grid files. Previously only Equilibrium options were
+  saved. Now also Mesh and nonorthogonal options (#98)\
+  By [John Omotani](https://github.com/johnomotani)
 - Setting to adjust extension of FineContours past targets, may help to avoid crashes on
   problematic equilibria (#96)\
   By [John Omotani](https://github.com/johnomotani)